requirements.txt
Flask==2.0.1
Werkzeug==2.0.2
flask-cors==3.0.10
gunicorn==20.1.0
requests==2.26.0
python-dotenv==0.19.0
supabase==0.7.1
pandas